// src/api/generated/model/getListingsParams.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * SkinTrades API
 * API for SkinTrades marketplace and trading platform. 

Features:
- **Catalog**: Browse global items.
- **Inventory**: Manage user items.
- **Marketplace**: Buy and sell items.
- **Trades**: P2P trading system.

Authentication:
Most endpoints require a valid JWT token. Use the `Authorize` button to enter your token.
 * OpenAPI spec version: v2.0
 */

export type GetListingsParams = {
search?: string;
weapon?: string;
rarity?: string;
type?: string;
team?: string;
page?: number;
size?: number;
};